Adds button to show VPC details to AWSWidget
To make this feature possible it was neccessary to add a feature which allows
to specify a so called userType when creating an AWS request. A userType is
a user-defined string which is added to a reply before the real reply data has
arrived. It allows to find out for which purpose the original request was made,
therefore allowing us to dispatch the reply to its intended destination even if
several AWS requests of the same type are used by a component.