Adds single instance protection
If SSH Commander is running and the user starts it again the new process
will terminate itself. Before it terminates it notifies the old process
to bring itself to the foreground.

This feature is per user. So several users can each run one instance of
SSH Commander on the same machine. It prevents multiple instances run by
the same user to race for writing their distinct JSON config files to
disk.