Fixes regression of macOS workaround for terminal redraw
If we hide a terminal which has no parent widget and make it visible again
Qt puts it into its own window.

This collides with the tunnel and SSHFS features which use a parentless
and therefore non-visible SSHTermWidget to control their background
processes of ssh and sshfs because the workaround makes those terminals
appear in their own window.

This fix disables the workaround for terminal widgets which have no
parent.